pyproject.toml
[project]
name = "dtocean-database"
version = "3.0.0"
description = "A database for storing persistent data, such as ocean energy converter (OEC) designs, metocean observations for sites, and reference data for electrical, mooring and foundations components, ports, vessels, etc."
authors = [
    {name = "Mathew Topper",email = "damm_horse@yahoo.co.uk"}
]
license = {text = "ODbL-1.0"}
readme = "README.md"
requires-python = ">=3.9,<3.13"

dependencies = [
    "dvc[s3] (>=3.59.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 80

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
